Add VersionFetcher class
