+++ packages/flutter_tools/gradle/src/main/kotlin/DependencyVersionChecker.kt
@@ -119,12 +119,12 @@ object DependencyVersionChecker {
     @JvmStatic fun checkDependencyVersions(project: Project) {
         project.extra.set(OUT_OF_SUPPORT_RANGE_PROPERTY, false)
 
-        checkGradleVersion(getGradleVersion(project), project)
-        checkJavaVersion(getJavaVersion(), project)
+        checkGradleVersion(VersionFetcher.getGradleVersion(project), project)
+        checkJavaVersion(VersionFetcher.getJavaVersion(), project)
 
         configureMinSdkCheck(project)
 
-        val agpVersion: AndroidPluginVersion? = getAGPVersion(project)
+        val agpVersion: AndroidPluginVersion? = VersionFetcher.getAGPVersion(project)
         if (agpVersion != null) {
             checkAGPVersion(agpVersion, project)
         } else {
@@ -134,7 +134,7 @@ object DependencyVersionChecker {
             )
         }
 
-        val kgpVersion: Version? = FlutterPluginUtils.getKGPVersion(project)
+        val kgpVersion: Version? = VersionFetcher.getKGPVersion(project)
         if (kgpVersion != null) {
             checkKGPVersion(kgpVersion, project)
         }
@@ -181,33 +181,12 @@ object DependencyVersionChecker {
         project: Project,
         it: Variant
     ): MinSdkVersion {
-        val agpVersion: AndroidPluginVersion? = getAGPVersion(project)
+        val agpVersion: AndroidPluginVersion? = VersionFetcher.getAGPVersion(project)
         return if (agpVersion != null && agpVersion.major >= 8 && agpVersion.minor >= 1) {
             MinSdkVersion(it.name, it.minSdk.apiLevel)
         } else {
             MinSdkVersion(it.name, it.minSdkVersion.apiLevel)
         }
-    }
-
-    // https://docs.gradle.org/current/kotlin-dsl/gradle/org.gradle.api.invocation/-gradle/index.html#-837060600%2FFunctions%2F-1793262594
-    @VisibleForTesting internal fun getGradleVersion(project: Project): Version {
-        val untrimmedGradleVersion: String = project.gradle.gradleVersion
-        // Trim to handle candidate gradle versions (example 7.6-rc-4). This means we treat all
-        // candidate versions of gradle as the same as their base version
-        // (i.e., "7.6"="7.6-rc-4").
-        return Version.fromString(untrimmedGradleVersion.substringBefore('-'))
-    }
-
-    // https://docs.gradle.org/current/kotlin-dsl/gradle/org.gradle.api/-java-version/index.html#-1790786897%2FFunctions%2F-1793262594
-    @VisibleForTesting internal fun getJavaVersion(): JavaVersion = JavaVersion.current()
-
-    @VisibleForTesting internal fun getAGPVersion(project: Project): AndroidPluginVersion? {
-        val androidPluginVersion: AndroidPluginVersion? =
-            project.extensions
-                .findByType(
-                    AndroidComponentsExtension::class.java
-                )?.pluginVersion
-        return androidPluginVersion
     }
 
     @VisibleForTesting internal fun getErrorMessage(
@@ -370,47 +349,6 @@ object DependencyVersionChecker {
             logger.error(warnMessage)
         }
     }
-}
-
-// Helper class to parse the versions that are provided as plain strings (Gradle, Kotlin) and
-// perform easy comparisons. All versions will have a major, minor, and patch value. These values
-// default to 0 when they are not provided or are otherwise unparseable.
-// For example the version strings "8.2", "8.2.2hfd", and "8.2.0" would parse to the same version.
-internal class Version(
-    val major: Int,
-    val minor: Int,
-    val patch: Int
-) : Comparable<Version> {
-    companion object {
-        fun fromString(version: String): Version {
-            val asList: List<String> = version.split(".")
-            val convertedToNumbers: List<Int> = asList.map { it.toIntOrNull() ?: 0 }
-            return Version(
-                major = convertedToNumbers.getOrElse(0) { 0 },
-                minor = convertedToNumbers.getOrElse(1) { 0 },
-                patch = convertedToNumbers.getOrElse(2) { 0 }
-            )
-        }
-    }
-
-    override fun compareTo(other: Version): Int {
-        if (major != other.major) {
-            return major - other.major
-        }
-        if (minor != other.minor) {
-            return minor - other.minor
-        }
-        if (patch != other.patch) {
-            return patch - other.patch
-        }
-        return 0
-    }
-
-    override fun equals(other: Any?): Boolean = other is Version && compareTo(other) == 0
-
-    override fun hashCode(): Int = major.hashCode() or minor.hashCode() or patch.hashCode()
-
-    override fun toString(): String = "$major.$minor.$patch"
 }
 
 // Custom error for when the dependency_version_checker.kts script finds a dependency out of
